Change in configuration class and project structure, to isolate configuration from application
