--- src/main/java/se/invidi/campaignbuilder/configuration/ApplicationConfiguration.java
@@ -1,10 +1,0 @@
-package se.invidi.campaignbuilder.configuration;
-
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-@ComponentScan(basePackages = "se.invidi")
-public class ApplicationConfiguration {
-
-}
